feat: implement getBorderPoints on mask
Refs: #121
